requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.0
idna==3.10
pymodbus==3.7.4
pyserial==3.5
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.2.3
